# docs/swagger.yaml
basePath: /
definitions:
  controllers.ChatRequest:
    properties:
      conversationId:
        description: 'Opcional: se não fornecido, cria nova conversa'
        type: string
      message:
        type: string
    required:
    - message
    type: object
  controllers.ChatResponse:
    properties:
      conversationId:
        type: string
      latencyMs:
        type: integer
      message:
        description: Resposta do assistente
        type: string
      messageId:
        type: string
      role:
        $ref: '#/definitions/models.MessageRole'
    type: object
...
  models.Conversation:
    properties:
      createdAt:
        type: string
      id:
        type: string
      title:
        description: Título da conversa (pode ser gerado automaticamente)
        type: string
      updatedAt:
        type: string
      userId:
        description: 'Opcional: para usuários autenticados'
        type: string
    type: object
...
  models.MessageRole:
    enum:
    - user
    - assistant
    - system
    type: string
    x-enum-varnames:
    - RoleUser
    - RoleAssistant
    - RoleSystem
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: API para chatbot SR Robot com JWT authentication e Prometheus metrics
  license:
    name: MIT
    url: http://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: SR Robot API
  version: "1.0"
paths:
  /api/v1/chat:
    post:
      consumes:
      - application/json
      description: Envia uma mensagem e recebe a resposta do chatbot. Cria nova conversa
        ou continua existente.
      parameters:
      - description: Mensagem do usuário
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.ChatRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.ChatResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Enviar mensagem para o chatbot
      tags:
      - chat
  /api/v1/conversations:
    get:
      consumes:
      - application/json
      description: Lista todas as conversas ordenadas por data de atualização
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Listar conversas
      tags:
      - chat
  /api/v1/conversations/{id}:
    delete:
      consumes:
      - application/json
      description: Deleta uma conversa e todas suas mensagens
      parameters:
      - description: Conversation ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Deletar conversa
      tags:
      - chat
    get:
      consumes:
      - application/json
      description: Retorna todas as mensagens de uma conversa específica
      parameters:
      - description: Conversation ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Obter histórico de conversa
      tags:
      - chat
    put:
      consumes:
      - application/json
      description: Atualiza o título de uma conversa existente
      parameters:
      - description: Conversation ID
        in: path
        name: id
        required: true
        type: string
      - description: Novo título
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Conversation'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Atualizar título da conversa
      tags:
      - chat
...
securityDefinitions:
  BearerAuth:
    description: Bearer token (add "Bearer " prefix)
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"